feature(category): 카테고리 캐싱 완료 (header, all)
