--- src/main/java/com/nhnacademy/back/product/category/service/UserCategoryService.java
@@ -1,13 +1,0 @@
-package com.nhnacademy.back.product.category.service;
-
-import java.util.List;
-
-import com.nhnacademy.back.product.category.domain.dto.response.ResponseCategoryDTO;
-
-public interface UserCategoryService {
-
-	List<ResponseCategoryDTO> getCategoriesToDepth3();
-
-	List<ResponseCategoryDTO> getCategoriesById(long categoryId);
-
-}
