+++ src/main/resources/application.yml
@@ -14,6 +14,9 @@ spring:
         dialect: org.hibernate.dialect.MySQL8Dialect # Hibernate 가 사용할 SQL 문법(방언)을 지정
     open-in-view: false          # 영속성 컨텍스트를 뷰까지 열지 않음
 
+  cache:
+    type: redis
+
   data:
     redis:
       host: s4.java21.net
